+++ packages/web/src/features/manage-club/members/frames/RegisterMemberListFrame.tsx
@@ -1,5 +1,7 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import styled from "styled-components";
+
+import { RegistrationDeadlineEnum } from "@sparcs-clubs/interface/common/enum/registration.enum";
 
 import AsyncBoundary from "@sparcs-clubs/web/common/components/AsyncBoundary";
 import FoldableSectionTitle from "@sparcs-clubs/web/common/components/FoldableSectionTitle";
@@ -9,7 +11,6 @@ import { useGetRegistrationTerm } from "@sparcs-clubs/web/features/clubs/service
 import useGetSemesterNow from "@sparcs-clubs/web/utils/getSemesterNow";
 
 import RegisterMemberList from "../components/RegisterMemberList";
-import { useRegistrationPeriod } from "../hooks/useRegistrationPeriod";
 
 const RegisterMemberListWrapper = styled.div`
   display: flex;
@@ -17,15 +18,39 @@ const RegisterMemberListWrapper = styled.div`
   gap: 20px;
 `;
 
-const RegisterMemberListFrame: React.FC = () => {
+const RegisterMemberListFrame = () => {
   const {
     data: termData,
     isLoading: isLoadingTerm,
     isError: isErrorTerm,
   } = useGetRegistrationTerm();
+  const [isRegistrationPeriod, setIsRegistrationPeriod] = useState<boolean>();
+  const [memberRegistrationPeriodEnd, setMemberRegistrationPeriodEnd] =
+    useState<Date>(new Date());
 
-  const { isRegistrationPeriod, memberRegistrationPeriodEnd } =
-    useRegistrationPeriod(termData ?? null);
+  useEffect(() => {
+    if (termData) {
+      const now = new Date();
+      const currentEvents = termData.events.filter(
+        event => now >= event.startTerm && now <= event.endTerm,
+      );
+      if (currentEvents.length === 0) {
+        setIsRegistrationPeriod(false);
+        return;
+      }
+      const registrationEvent = currentEvents.filter(
+        event =>
+          event.registrationEventEnumId ===
+          RegistrationDeadlineEnum.StudentRegistrationApplication,
+      );
+      if (registrationEvent.length > 0) {
+        setIsRegistrationPeriod(true);
+        setMemberRegistrationPeriodEnd(registrationEvent[0].endTerm);
+      } else {
+        setIsRegistrationPeriod(false);
+      }
+    }
+  }, [termData]);
 
   const {
     semester: semesterInfo,
